# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Local Cloud API",
    "version": "1.0.0",
    "description": "Local Cloud API for the Nextcloud alternative."
  },
  "servers": [
    {
      "url": "http://localhost:4170",
      "description": "Default local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/file": {
      "get": {
        "summary": "Stream file",
        "parameters": [
          {
            "name": "root",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "download",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stream"
          }
        }
      }
    },
    "/api/preview": {
      "get": {
        "summary": "Image/video preview",
        "parameters": [
          {
            "name": "root",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Preview"
          }
        }
      }
    },
    "/api/album-art": {
      "get": {
        "summary": "Album art",
        "parameters": [
          {
            "name": "root",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Art"
          }
        }
      }
    },
...
  }
}